feat(organisms): move BuyFlow to organisms folder
