+++ src/components/organisms/Buyflow/index.tsx
@@ -1,8 +1,8 @@
 import React, { useState } from "react";
-import AgeStep from "../molecules/AgeStep";
-import SummaryStep from "../molecules/SummaryStep";
-import EmailStep from "../molecules/EmailStep";
-import { BuyflowProps, PRODUCT_IDS_TO_NAMES } from "../types";
+import AgeStep from "../../molecules/AgeStep";
+import SummaryStep from "../../molecules/SummaryStep";
+import EmailStep from "../../molecules/EmailStep";
+import { BuyflowProps, PRODUCT_IDS_TO_NAMES } from "../../types";
