+++ src/pages/buy/insurance_dev.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Buyflow from "@/components/buyflow/Buyflow";
+import Buyflow from "@/components/organisms/Buyflow";
 import { ProductIds } from "@/components/types";
 
 export default function InsuranceDev() {
